refactor: move device listing to static method on `LoupedeckDevice`
