+++ index.js
@@ -1,9 +1,7 @@
 const LoupedeckDevice = require('./device')
 const { HAPTIC } = require('./constants')
-const listDevices = require('./discover')
 
 module.exports = {
     LoupedeckDevice,
-    HAPTIC,
-    listDevices
+    HAPTIC
 }
